pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bud-rag"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "faiss-cpu>=1.7",
    "numpy>=1.24",
    "mcp>=1.1.0",
    "prompt-toolkit>=3.0",
    "anthropic>=0.86.0",
]

[project.scripts]
bud = "bud.cli:main"
bud-mcp = "bud.bud_mcp:main"
buddy = "buddy.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
]

[tool.hatch.build.targets.wheel]
packages = ["bud", "buddy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
